requirements.txt
beautifulsoup4==4.12.3
matplotlib==3.9.1
moviepy==1.0.3
numpy==2.0.1
pandas==2.2.2
Pillow==9.5.0
selenium==4.23.1
